fix(opcodes): cast fprintf args to match %X under -Wformat -Werror
The invalid-op diagnostic prints the opcode and program counter using %04X.
In varargs calls, `uint16_t` operands are promoted to `int`, so `c8->pc - 2`
was being passed as an `int` while the format string expects `unsigned int`.
With -Wformat enabled and promoted via -Werror, this caused a hard failure:

  format `%X` expects argument of type `unsigned int`, but argument has type `int`
  [-Werror=format=]

Rewrite the fprintf call to explicitly cast the opcode and PC expression to
`unsigned int` (and keep the PC decrement typed) so the format contract is
satisfied and strict builds remain clean.

+++ src/opcodes.c
@@ -100,7 +100,10 @@ static Chip8Op tableF[256] = {
 static inline void
 op_invalid(Chip8 *c8, uint16_t op)
 {
-	fprintf(stderr, "Invalid op %04X at PC=%04X\n", op, c8->pc - 2);
+	fprintf(stderr,
+		"Invalid op %04X at PC=%04X\n",
+		(unsigned int)op,
+		(unsigned int)((uint16_t)(c8->pc - 2u)));
 }
 
 /* ============================================================
